requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-import-export==3.0.2
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
